General cleanup (#35)
* Format Code and improve formatting

- Formatted the code
- Fixed the `pyproject.toml` on the backend so that it actually reformats code when we run a command like `uv run ruff format`

* Updated package.json when running dev and preview

Updated package.json when running dev and preview. Now it explicitly runs it with a host and port. The main part of this was running it with host `127.0.0.1`, which aligns with our configurations of detecting when we're running something in Development server.'

* Removed remaining H2S script files

Had some remaining H2S related files that were related to how we tried to do client side
calculations for H2S. We've moved those calculations to the server side so we don't need these files anymore.

* Prevented archived users from being authenticated

- Fixed makefile as well